entity_graph_broken

+++ src/main/java/com/example/demo/services/CarService.java
@@ -1,20 +1,23 @@
 package com.example.demo.services;
 
+import com.example.demo.controllers.CarController;
 import com.example.demo.entities.CarEntity;
 import com.example.demo.entities.LocationEntity;
-import com.example.demo.enums.Category;
-import com.example.demo.enums.Transmission;
+import com.example.demo.error.CarExistsException;
+import com.example.demo.error.IdNotFoundException;
 import com.example.demo.repositories.CarRepository;
 import com.example.demo.repositories.LocationRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.hateoas.mvc.ControllerLinkBuilder;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
-import org.springframework.web.server.ResponseStatusException;
 
+import java.net.URI;
+import java.util.List;
 import java.util.Optional;
 
 @Service
@@ -22,50 +25,36 @@ public class CarService implements ICarService {
     @Autowired CarRepository carRepository;
     @Autowired LocationRepository locationRepository;
 
-    public CarEntity createCar (String name, String registration, String manufacturer,
-                                String model, String transmission, String category,
-                                long location_id) {
-
-        Category car_category = null;
-        Transmission car_transmission = null;
-
-        for (Category cat : Category.values()){
-            if (cat.toString().equalsIgnoreCase(category)){
-                car_category = cat;
-            }
-        }
-
-        for (Transmission tran : Transmission.values()){
-            if (tran.toString().equalsIgnoreCase(transmission)){
-                car_transmission = tran;
-            }
-        }
-
-        CarEntity newCar = new CarEntity(name, registration, manufacturer, model, car_transmission, car_category, locationRepository.findById(location_id).get());
-
-
-        if (car_category == null){
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "invalid category.");
-        }
-        else if (car_transmission == null){
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "invalid transmission.");
+    @Override public ResponseEntity createCar(CarEntity car) {
+        CarEntity addCar = carRepository.findOneByRegistration(car.getRegistration());
+        if (addCar == null) {
+            carRepository.save(car);
+            URI u = ControllerLinkBuilder.linkTo(ControllerLinkBuilder.methodOn(CarController.class).findCustomer(car.getId())).toUri();
+            return ResponseEntity.ok(u);
         } else{
-            return new ResponseEntity<>(newCar, headers, HttpStatus.CREATED);
+            throw new CarExistsException(car.getRegistration());
         }
-
-
-        carRepository.save(newCar);
-        return newCar;
     }
 
     @Override
-    public void deleteCar (Long id) {
-        carRepository.deleteById(id);
+    public ResponseEntity deleteCar (Long id) {
+        Optional<CarEntity> deleteCar = carRepository.findById(id);
+        if (deleteCar.isPresent()){
+            carRepository.deleteById(id);
+            return new ResponseEntity("car deleted successfully.", new HttpHeaders(), HttpStatus.NO_CONTENT);
+        } else {
+            throw new IdNotFoundException(id);
+        }
     }
 
     @Override
-    public Optional<CarEntity> findById (Long id) {
-        return carRepository.findById(id);
+    public ResponseEntity findById (Long id) {
+        Optional<CarEntity> foundCar = carRepository.findById(id);
+        if (foundCar.isPresent()){
+            return new ResponseEntity(foundCar.get(), HttpStatus.OK);
+        } else{
+            throw new IdNotFoundException(id);
+        }
     }
 
     @Override
@@ -77,4 +66,20 @@ public class CarService implements ICarService {
     public Page<CarEntity> findAll (Pageable pageable) {
         return carRepository.findAll(pageable);
     }
+
+    @Override
+    public List<CarEntity> findByLocation(Long location) {
+        LocationEntity loc = locationRepository.findById(location).get();
+        return carRepository.findAllByLocation(loc);
+
+        /*
+        Optional<LocationEntity> loc = locationRepository.findById(location);
+        if (loc.isPresent()){
+            return new ResponseEntity(carRepository.findAllByLocation(loc.get()), HttpStatus.OK);
+        } else{
+            throw new IdNotFoundException(location);
+        }
+         */
+    }
+
 }
